Cleaned up NodeAttr type concept by adding key
This work was mainly focused on clarifying 'type' field to reference a C data type,
and a new 'key' field for arbitrary field names (as enum values).

This change allowed me to remove the super ugly enum hack for supporting
functions and generalize that out to a specified type with arbitrary
key (names).

Function attributes (and other attrs) can now be applied with the
new new_ext_ptr_attr() factory. This is used for any pointer attribute
that is externally owned and should not be freed from free_attr calls.

Also fixed up some lingering (and new) valgrind errors, so valgrind
reports a clean build.