Removed global arena and global allocation methods
Still have a global (module-scope) allocator, but it's in the node
module now, so it's not as likely to cause confusion across a variety of
use cases or modules. That allocator must be provided to the node
implementation before attempting to use runtime allocations.

Added setup and teardown features to the test_helper to avoid too much
duplication throughout the test infrastructure.